% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22081.R
\name{regres}
\alias{regres}
\title{self coded regression}
\usage{
regres(df, method, lam = 0, k = 1, testpoint = 0)
}
\arguments{
\item{df}{data frame containing the data of predicting variable as well as predicted variable,noted that the predicted variable has to be 
at the last column of the given data frame.}

\item{method}{0 to preform standard regression,1 for ridge regression and 2 for locally weighted regression}

\item{lam}{lambda the multiple of identity matrix added to xTx in the process of ridge regression}

\item{k}{the constant needed in calculating the Gaussian kernel for locally weighted linear regression.}

\item{testpoint}{the point chosen for the locally weighted regression}
}
\value{
the coefficient of the regression whose last column is the value of intercept
}
\description{
apply standard,ridge or locally weighted regression(using Gaussian kernel) on a given data frame(the predicted variable has to be at the last column 
of the data frame like 'Boston' in MASS we used in vignettes )
}
\examples{
\dontrun{
coef <- regres(df, 1, lam = 0.2)
}
}
